Refactor BQ stream source to use blocks.
This change modifies the BigQueryStorageStreamReader to inherit
from a class that looks  a lot like BlockBasedReader by duplicating
many of the interfaces found in that reader and, to a lesser
extent, in FileBasedReader. Ideally we would avoid this
duplication either by refactoring the inheritance structure of
the BlockBasedSource/BlockBasedReader so that we could inherit from
this type without depending on FileBasedReader or by finding a way
to expose the GAX ServerStream object as a ReadableByteChannel.