Report estimated status in BQ stream reader.
This change modifies the BigQueryStorageStreamReader object to
record the estimated stream length as reported by the server and
to provide progress estimates to the Beam runner using the
estimated stream length value. Since the reader does not support
splits at the moment (splitAtFraction is not implemented), there is
no need to synchronize access to the source and the estimated
stream length value, but this will become necessary when and if we
support stream splits.