Range tracker for BQ stream offsets
This change creates a new StreamOffsetRangeTracker which implements
functionality specific to the BigQueryStorageStreamReader. In
particular, it supports the notion of pending splits, where the
state of the reader can be updated from within splitAtFraction()
to indicate that a split is pending, and any further attempts to
return a row will fail until the start() / advance() thread has a
chance to kill the existing server stream, call the server-side
split API, and create a new read session in the context of the new
primary or residual stream.