Use an in-memory gRPC server in test
This change modifies the BigQueryServices class to expose the client
object for the BigQuery storage service directly; this allows the stream
reader to keep a ServerStream object, and to cancel the stream if the
reader is cancelled. To improve test coverage, this change also removes
the fake table read service and instead starts an in-memory server which
implements the BigQuery storage service interface, and uses real server
stream objects when testing.