Add Avro support to BigQuery stream sources.
This change modifies the BigQuery storage stream source to support
reading in Avro format in addition to the existing Row proto format.
It's likely that, for Beta, Avro will be the only supported format for
the read API and that we'll be deprecating the Row proto format
entirely; however, I've gone to the trouble of refactoring into a base
class and format-specific subclasses since we will at some point add
public support for Apache Arrow and potentially other popular third-
party formats.