Pass TableSchema when calling Avro parseFns.
This change modifies the BigQuery storage stream source for Avro to
store the table schema for the table when creating a stream source and
to pass the schema as part of the SchemaAndRecord object supplied for
each invocation of the stream parse function. This should allow existing
pipelines, including those authored against the legacy TableRow type, to
be moved to the read API transparently.

Note there is the potential for a small race condition in which the
schema of the table can change between when the table schema is fetched
and when the read session is created; this is no worse than what we have
today, since the same race condition exists where the export job is
invoked, but it's still there.